+++ installer/frontend/components/nodes.jsx
@@ -4,23 +4,28 @@ import { connect } from 'react-redux';
 
 import {
   AWS_CONTROLLERS,
+  AWS_ETCDS,
   AWS_WORKERS,
+  ETCD_OPTION,
+  ETCD_OPTIONS,
+  EXTERNAL_ETCD_CLIENT,
   IAM_ROLE,
   IAM_ROLE_CREATE_OPTION,
   INSTANCE_TYPE,
   NUMBER_OF_INSTANCES,
+  PLATFORM_TYPE,
   STORAGE_IOPS,
   STORAGE_SIZE_IN_GIB,
   STORAGE_TYPE,
 } from '../cluster-config';
 
-import { Form } from '../form';
-import { toError } from '../utils';
-import { validate } from '../validate';
 import { AWS_INSTANCE_TYPES } from '../facts';
-import { NumberInput, Connect, Select } from './ui';
+import { Field, Form } from '../form';
+import { AWS_TF } from '../platforms';
+import { toError } from '../utils';
+import { compose, validate } from '../validate';
 import { makeNodeForm, toKey } from './make-node-form';
-import { Etcd } from './etcd';
+import { Connect, Input, NumberInput, Radio, Select } from './ui';
 
 const Row = ({label, htmlFor, children}) => <div className="row form-group">
   <div className="col-xs-4">
@@ -103,15 +108,103 @@ export const DefineNode = ({type, max, withIamRole = true}) => <div>
 const MAX_MASTERS = 10;
 const MAX_WORKERS = 1000;
 
-// TODO (kans): add ectdForm here
-const fields = [
+const etcdForm = new Form('etcdForm', [
+  new Field(ETCD_OPTION, {default: ETCD_OPTIONS.PROVISIONED}),
+  makeNodeForm(AWS_ETCDS, compose(validate.int({min: 1, max: 9}), validate.isOdd), false, {
+    dependencies: [ETCD_OPTION],
+    ignoreWhen: cc => cc[ETCD_OPTION] !== ETCD_OPTIONS.PROVISIONED,
+  }),
+  new Field(EXTERNAL_ETCD_CLIENT, {
+    default: '',
+    validator: validate.url,
+    dependencies: [ETCD_OPTION],
+    ignoreWhen: cc => cc[ETCD_OPTION] !== ETCD_OPTIONS.EXTERNAL,
+  }),
+], {
+  validator: value => {
+    const etcd = value[ETCD_OPTION];
+    if (!_.values(ETCD_OPTIONS).includes(etcd)) {
+      return 'Please select an option.';
+    }
+  },
+});
+
+const awsNodesForm = new Form('awsNodesForm', [
   makeNodeForm(AWS_CONTROLLERS, validate.int({min: 1, max: MAX_MASTERS})),
   makeNodeForm(AWS_WORKERS, validate.int({min: 1, max: MAX_WORKERS})),
-];
+  etcdForm,
+]);
 
-const form = new Form('DefineNodesForm', fields);
+export const Etcd = connect(({clusterConfig}) => ({
+  etcdOption: clusterConfig[ETCD_OPTION],
+  isAWS: clusterConfig[PLATFORM_TYPE] === AWS_TF,
+}))(
+  ({etcdOption, isAWS}) => <div>
+    <div className="row form-group">
+      <div className="col-xs-12">
+        etcd is the key-value store used by Kubernetes for cluster coordination and state management.
+      </div>
+    </div>
 
-export const AWS_DefineNodes = () => <div>
+    <div className="row form-group">
+      <div className="col-xs-12">
+        <div className="wiz-radio-group">
+          <div className="radio wiz-radio-group__radio">
+            <label>
+              <Connect field={ETCD_OPTION}>
+                <Radio name={ETCD_OPTION} value={ETCD_OPTIONS.PROVISIONED} id={ETCD_OPTIONS.PROVISIONED} />
+              </Connect>
+              {isAWS && <span>Provision AWS etcd cluster</span>}
+              {!isAWS && <span>Provision etcd cluster directly on controller nodes</span>}
+            </label>&nbsp;(default)
+            <p className="text-muted wiz-help-text">
+              {isAWS && <span>Create EC2 instances to run an etcd cluster.</span>}
+              {!isAWS && <span>Run etcd directly on controller nodes.</span>}
+            </p>
+          </div>
+          <div className="radio wiz-radio-group__radio">
+            <label>
+              <Connect field={ETCD_OPTION}>
+                <Radio name={ETCD_OPTION} value={ETCD_OPTIONS.EXTERNAL} id={ETCD_OPTIONS.EXTERNAL} />
+              </Connect>
+              I have my own etcd v3 cluster
+            </label>
+            <p className="text-muted wiz-help-text">Your Tectonic cluster will be configured to use an external etcd, which you specify.</p>
+          </div>
+        </div>
+        <etcdForm.Errors />
+      </div>
+    </div>
+    {etcdOption === ETCD_OPTIONS.EXTERNAL && <hr />}
+    {etcdOption === ETCD_OPTIONS.EXTERNAL &&
+      <div className="form-group">
+        <div className="row">
+          <div className="col-xs-3">
+            <label htmlFor={EXTERNAL_ETCD_CLIENT}>Client Address</label>
+          </div>
+          <div className="col-xs-8">
+            <Connect field={EXTERNAL_ETCD_CLIENT}>
+              <Input id={EXTERNAL_ETCD_CLIENT}
+                autoFocus={true}
+                className="wiz-inline-field wiz-inline-field--suffix"
+                suffix={<span className="input__suffix">:2379</span>}
+                placeholder="https://etcd.example.com" />
+            </Connect>
+            <p className="text-muted">Address of etcd client endpoint</p>
+          </div>
+        </div>
+      </div>
+    }
+    {isAWS && etcdOption === ETCD_OPTIONS.PROVISIONED && <hr />}
+    {isAWS && etcdOption === ETCD_OPTIONS.PROVISIONED &&
+      <div className="row form-group col-xs-12">
+        <DefineNode type={AWS_ETCDS} max={9} withIamRole={false} />
+      </div>
+    }
+  </div>
+);
+
+export const AWS_Nodes = () => <div>
   <h3>Master Nodes</h3>
   <br />
   <DefineNode type={AWS_CONTROLLERS} max={MAX_MASTERS} />
@@ -125,4 +218,5 @@ export const AWS_DefineNodes = () => <div>
   <Etcd />
 </div>;
 
-AWS_DefineNodes.canNavigateForward = state => form.canNavigateForward(state) && Etcd.canNavigateForward(state);
+Etcd.canNavigateForward = etcdForm.canNavigateForward;
+AWS_Nodes.canNavigateForward = awsNodesForm.canNavigateForward;
